Fixed chaos spawn times
I changed how it checks the difficulty because the order it runs the files made it not possible to check the difficulty the way it was before. This should make chaos what it was supposed to be. Also changed the number to 7 because i believe the numbers got changed so a difficulty wasn't set to 0