Set numbers more accordingly
I plan on keeping the difficulty variable for enemy count available but mostly unused to keep balancing each difficulty much easier. With a consistent amount of enemies it will be easier to adjust health and damage for them on a difficulty basis. I also removed a "coefficient" variable that isn't used anywhere else