Skefil's QOL Stuff Part 3: Winchester LAR
Part 3 of replacing some gun models with better ones for quality of life. This one replaces the Winchester LAR to use the Model 1873 from M9K/CS+. Stats and performance of the original weapon have been carried over to the new one. The gun also uses a skin that matches closer to the 1887 VM it's replacing.